Organize assets into folders
