+++ css/style.css
@@ -10,7 +10,7 @@
       display: block;
       width: 100vw;
       height: 100vh;
-      background: url('background.png') no-repeat center center / cover;
+      background: url('../images/background.png') no-repeat center center / cover;
     }
     #hud {
       position: absolute;
