+++ index.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Mage Roguelike Prototype</title>
   
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="css/style.css">
 
 </head>
 <body>
@@ -55,11 +55,11 @@
       <button id="btnE">E</button>
     </div>
   </div>
-  <script src="constants.js"></script>
-  <script src="spawner.js"></script>
-  <script src="upgrades.js"></script>
-  <script src="renderer.js"></script>
-  <script src="script.js"></script>
+  <script src="js/constants.js"></script>
+  <script src="js/spawner.js"></script>
+  <script src="js/upgrades.js"></script>
+  <script src="js/renderer.js"></script>
+  <script src="js/script.js"></script>
 
 </body>
 </html>
